Separate waveFunctionCollapse.py into several files for ease of use, readability and expansion. The original waveFunctionCollapse.py file was refactored into: - initiatilzation.py - encapsulates the initialization phase of the program (creation of rules set, initial wave matrix, extraction of patterns etc.). - iteration.py - includes the functionalities of the main loop of the algorithm (observe, collapse, propagate). - rendering.py - handles all rendering matters of the code (saving images from iterations, creating the movie etc.). - main.py - include the main wave function collapse function, which preforms the initiation, iterations, and rendering. - waveFunctionCollapse.py - handles user commandline input parsing and validation, then calls the main function.
Note that this commit does not change any functionality of the code itself.
